Extract storage-channels module from wall profiler
Move the dd-trace storage diagnostics channels (storage:enter,
storage:before, span:finish, span:tags:update), the legacy-storage
enterWith/run shimmer, and the getActiveSpan helper out of
profiling/profilers/wall.js into a new packages/dd-trace/src/storage-channels.js.

No behavior change for the wall profiler. The extraction is to let a
forthcoming OTEP-4947 thread-context writer reuse the same channel
infrastructure without duplicating the shimmer or pulling in the
profiler module.

Adjust wall.spec.js's proxyquire stubs to replace the new module
instead of mocking datadog-core directly.
