+++ .github/CODEOWNERS
@@ -370,6 +370,7 @@
 
 /packages/dd-trace/*/profiling/ @DataDog/profiling-js
 /packages/dd-trace/src/profiler.js @DataDog/profiling-js
+/packages/dd-trace/src/storage-channels.js @DataDog/profiling-js @DataDog/apm-sdk-capabilities-js
 /packages/dd-trace/test/exporters/common/form-data.spec.js @DataDog/profiling-js
 
 # Language Platform
