+++ packages/dd-trace/test/profiling/profilers/wall.spec.js
@@ -470,12 +470,13 @@ describe('profilers/native/wall', () => {
 
       WallProfiler = proxyquire('../../../src/profiling/profilers/wall', {
         '@datadog/pprof': localPprof,
-        '../../../../datadog-core': {
-          storage: () => ({
-            getStore: () => currentStore,
-            enterWith () {},
-            run (store, cb, ...args) { return cb(...args) },
-          }),
+        '../../storage-channels': {
+          enterCh,
+          beforeCh: dc.channel('dd-trace:storage:before'),
+          spanFinishCh: dc.channel('dd-trace:span:finish'),
+          tagsUpdateCh: dc.channel('dd-trace:span:tags:update'),
+          getActiveSpan: () => currentStore && currentStore.span,
+          ensureChannelsActivated: () => {},
         },
       })
     })
@@ -774,12 +775,13 @@ describe('profilers/native/wall', () => {
 
       WallProfiler = proxyquire('../../../src/profiling/profilers/wall', {
         '@datadog/pprof': localPprof,
-        '../../../../datadog-core': {
-          storage: () => ({
-            getStore: () => currentStore,
-            enterWith () {},
-            run (store, cb, ...args) { return cb(...args) },
-          }),
+        '../../storage-channels': {
+          enterCh,
+          beforeCh: dc.channel('dd-trace:storage:before'),
+          spanFinishCh: dc.channel('dd-trace:span:finish'),
+          tagsUpdateCh: dc.channel('dd-trace:span:tags:update'),
+          getActiveSpan: () => currentStore && currentStore.span,
+          ensureChannelsActivated: () => {},
         },
       })
     })
